Task is now a kind of activity instead of an arc.
A reason is to have a clearer separation between pure activities and
engine-controlled tasks. Before the change, an activity could do both a pure
computation and also fire a task. Another reason is to make it possible in the
future to have activities result in multiple tokens, but not in multiple tasks.